requirements.txt
tweepy>=4.14.0
instagrapi>=2.0.0
atproto>=0.0.54
Pillow>=10.0.0
requests>=2.31.0